requirements.txt
yfinance
pandas
pyyaml
python-dotenv
matplotlib
streamlit
openai>=1.40.0
